workers/requirements.txt
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.5.0
schedule==1.2.0
psycopg2-binary==2.9.9
